add update and delete jajanan

+++ database/migrations/2024_03_18_023619_create_jajanans_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('pedagang_id');
             $table->string('nama');
             $table->string('deskripsi');
-            $table->float('harga');
+            $table->integer('harga');
             $table->string('image')->nullable()->default(null);
             $table->boolean('tersedia')->default(true);
             $table->string('kategori');
